+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+* consolidate signals: in `src/components/info.rs`
 * robust error handling: in `info.rs`, `use_future`
 * `command_host()` param: the ip is validated
 * refactoring: UI code in `info.rs`